raw conversion of AssignsOutAndRefParameters specs
